requirements.txt
flask
beautifulsoup4
opencv-python
tensorflow==1.12.0
Pillow